backend/requirements.dev.txt
django
djangorestframework
dj-database-url
psycopg-binary
python-decouple
requests
djangorestframework-simplejwt[crypto]
django-cors-headers